git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@1877 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/mixer/read_frame.h
@@ -41,12 +41,12 @@ class read_frame : boost::noncopyable
 {
 public:
 	read_frame();
-	read_frame(const safe_ptr<ogl_device>& ogl, size_t size, safe_ptr<host_buffer>&& image_data, audio_buffer&& audio_data);
+	read_frame(const safe_ptr<ogl_device>& ogl, int size, safe_ptr<host_buffer>&& image_data, audio_buffer&& audio_data);
 
 	virtual const boost::iterator_range<const uint8_t*> image_data();
 	virtual const boost::iterator_range<const int32_t*> audio_data();
 
-	virtual size_t image_size() const;
+	virtual int image_size() const;
 		
 private:
 	struct implementation;
