+++ core/producer/stage.h
@@ -61,8 +61,8 @@ public:
 	void clear(int index);
 	void clear();	
 	void swap_layers(const safe_ptr<stage>& other);
-	void swap_layer(int index, size_t other_index);
-	void swap_layer(int index, size_t other_index, const safe_ptr<stage>& other);
+	void swap_layer(int index, int other_index);
+	void swap_layer(int index, int other_index, const safe_ptr<stage>& other);
 	
 	boost::unique_future<std::wstring>				call(int index, bool foreground, const std::wstring& param);
 	boost::unique_future<safe_ptr<frame_producer>>	foreground(int index);
